testing alarm with repetead days

+++ app/src/main/java/com/example/noteapp/presentation/alarm/AlarmScheduler.kt
@@ -42,7 +42,7 @@ class AlarmScheduler @Inject constructor(private val context: Context) {
 
         val pendingIntent = PendingIntent.getBroadcast(
             context,
-            note.id!!.toInt(),
+            note.id!!,
             intent,
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
         )
